+++ docs/changelog.md
@@ -34,6 +34,10 @@
 
 ## 更新说明
 
+1.7.77 fix: fix index_value_hist_funddb interface
+
+    1. 修复 index_value_hist_funddb 接口
+
 1.7.76 fix: fix macro_euro_gdp_yoy interface
 
     1. 修复 macro_euro_gdp_yoy 接口
@@ -1415,6 +1419,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.7.77 fix: fix index_value_hist_funddb interface
 
 1.7.76 fix: fix macro_euro_gdp_yoy interface